Upgrade create_calendar_event and refresh behaviour of UNGTestMixin
Method create_calendar_event now support both start and end properties
of month, day and year.
It was also added the possibility to disable the behaviour "refresh untill
see event on calendar", to support the creation of events with
a date different from that selected under calendar to be shown.